Kernelloader 1.7 - Enabled screenshot feature. Screenshot in kernelloader can be done by   pressing button "R1". Screenshot is saved on USB memory stick. - Allow only one RPC call at the same time. This fixed CDVD bug. - Allow to set graphic mode. Implementation of function setdve is still missing. - Fixed memory hole when SifSetDma() is currently not possible in   SifCallRpc() and SifbindRpc(). - Cleanup return codes when SBIOS call need to be retried later.

+++ TGE/sbios/sifcmd.c
@@ -206,7 +206,7 @@ static void sif_cmd_interrupt()
 
 	if ((cmd_handlers != NULL) && (cmd_handlers[id].handler != NULL)) {
 #ifdef SBIOS_DEBUG
-		printf("Callback 0x%x\n", cmd_handlers[id].handler);
+		printf("Callback 0x%x\n", (uint32_t) cmd_handlers[id].handler);
 #endif
 		cmd_handlers[id].handler(packet, cmd_handlers[id].harg);
 	}
